// frontend/src/services/generated-zod/answer-controller/answer-controller.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import {
  z as zod
} from 'zod'

/**
 * Spočítá odměnu, uloží auditní záznam odpovědi a aktualizuje score hráče.
 * @summary Vyhodnotit odpověď hráče
 */
export const submitAnswerBodySessionIdMin = 0;

export const submitAnswerBodySessionIdMax = 64;
export const submitAnswerBodyRemainTimeMin = 0;

export const submitAnswerBodyRemainTimeMax = 60;


export const submitAnswerBody = zod.object({
  "player_id": zod.number(),
  "question_id": zod.number(),
  "session_id": zod.string().min(submitAnswerBodySessionIdMin).max(submitAnswerBodySessionIdMax),
  "is_phishing": zod.boolean(),
  "remain_time": zod.number().min(submitAnswerBodyRemainTimeMin).max(submitAnswerBodyRemainTimeMax)
})